+++ tests/unit/providers/test_injections_py2_py3.py
@@ -13,7 +13,7 @@ class PositionalInjectionTests(unittest.TestCase):
 
     def test_get_value_with_not_provider(self):
         injection = providers.PositionalInjection(123)
-        self.assertEquals(injection.get_value(), 123)
+        self.assertEqual(injection.get_value(), 123)
 
     def test_get_value_with_factory(self):
         injection = providers.PositionalInjection(providers.Factory(object))
@@ -61,11 +61,11 @@ class NamedInjectionTests(unittest.TestCase):
 
     def test_get_name(self):
         injection = providers.NamedInjection('name', 123)
-        self.assertEquals(injection.get_name(), 'name')
+        self.assertEqual(injection.get_name(), 'name')
 
     def test_get_value_with_not_provider(self):
         injection = providers.NamedInjection('name', 123)
-        self.assertEquals(injection.get_value(), 123)
+        self.assertEqual(injection.get_value(), 123)
 
     def test_get_value_with_factory(self):
         injection = providers.NamedInjection('name',