+++ docs/main/changelog.rst
@@ -12,6 +12,7 @@ Development version
 - Fix bug `#208 <https://github.com/ets-labs/python-dependency-injector/issues/208>`_:
   version ``3.14.0`` hasn't worked on Python 3.5.2 (thanks to
   `Jeroen Entjes <https://github.com/JeroenEntjes>`_).
+- Remove deprecated ``assertEquals`` from tests.
 
 3.14.0
 ------